+++ android/src/main/java/com/vydia/RNUploader/files/helpers/FilesHelperImpl.kt
@@ -1,4 +1,4 @@
-package com.vydia.RNUploader.files
+package com.vydia.RNUploader.files.helpers
 
 import java.io.File
