+++ android/src/main/java/com/vydia/RNUploader/files/helpers/MimeTypeHelper.kt
@@ -1,4 +1,4 @@
-package com.vydia.RNUploader.files
+package com.vydia.RNUploader.files.helpers
 
 interface MimeTypeHelper {
